Create a new RHIMemoryStatisticsInterface and expose all mem stats functions through that interface. Move mem stats to/from json util functions to RHIUtils
